Hack System Prototype (change: events/delgations)
I've gone through a video and with the help of ai changed the before prototype to be with events/delegations, i have in the works the fleshing out of the prototype in a short few days (1-2) to flesh a lot of things out and make it more presentable and to a standard of what is excepted.

Not much has changed interns other than this move.